requirements.txt
absl-py
aiodns
aiohappyeyeballs
aiohttp
aiosignal
annotated-doc
annotated-types
anyio
argcomplete
attrs
av>=14.0.0
beautifulsoup4
borb
certifi
cffi
charset-normalizer
click
contourpy
cryptography
curlify
cycler
datasets
dill
filelock
flatbuffers
fonttools
frozenlist
fsspec
gdown
gitdb
gitpython
h11
hf-xet
httpcore
httpx
huggingface-hub
idna
imageio
jinja2
joblib
kiwisolver
lazy-loader
lightning-utilities
lxml
markdown-it-py
markupsafe
matplotlib
mdurl
mediapipe
mpmath
multidict
multiprocess
networkx
numpy
opencv-contrib-python
opencv-python
pandas
pillow
platformdirs
propcache
protobuf
psutil
pyarrow
pycares
pycparser
pycryptodome
pydantic
pydantic-core
pygments
pyparsing
pysocks
python-barcode
python-dateutil
pytorch-lightning
pyyaml
qrcode
requests
rich
scikit-image
scikit-learn
scipy
sentencepiece
sentry-sdk
setuptools
shellingham
six
smmap
sounddevice
soupsieve
sympy
threadpoolctl
tifffile
torchmetrics
tqdm
typer
typing-extensions
typing-inspection
urllib3
vastai
wandb
xdg
xxhash
yarl
kornia